GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   branch head: d3e59acbb3279ff92208510c731ebd9661f71066
   - .vscode/extensions.json
   - .gitignore

Your previous branch was: refs/heads/main

The sha for that commit was: 4bbab66aa42dffdc6453ff257ff46379841dece1

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch